Use links array with map function to add the links in the sidebar instead of using several <Li> components
